# server: speed up single gguf creates