--- a/server/model.go
+++ b/server/model.go
@@ -180,10 +180,10 @@ func parseFromFile(ctx context.Context, file *os.File, digest string, fn func(ap
 		if digest != "" && n == stat.Size() && offset == 0 {
 			layer, err = NewLayerFromLayer(digest, mediatype, file.Name())
 			if err != nil {
-				slog.Debug("could not create new layer from layer", "error", err)			
+				slog.Debug("could not create new layer from layer", "error", err)
 			}
-		} 
-		
+		}
+
 		// Fallback to creating layer from file copy (either NewLayerFromLayer failed, or digest empty/n != stat.Size())
 		if layer == nil {
 			layer, err = NewLayer(io.NewSectionReader(file, offset, n), mediatype)
@@ -191,7 +191,6 @@ func parseFromFile(ctx context.Context, file *os.File, digest string, fn func(ap
 				return nil, err
 			}
 		}
-		
 
 		layers = append(layers, &layerGGML{layer, ggml})
 		offset = n