the blob for a file with a single gguf is already copied to the server on `/api/blobs/:digest`.
on `createModel`, we can avoid the rewriting of this blob.

This does not improve create speeds for safetensors or files with multiple gguf files

New logs
```
[GIN] 2024/07/23 - 17:21:45 | 201 |  5.298126708s |       127.0.0.1 | POST     "/api/blobs/sha256:54696cbcadd1959275fc99f9cc67880d2f38419124da06cdf2140bad2dc3d94c"
[GIN] 2024/07/23 - 17:21:45 | 200 |    8.519125ms |       127.0.0.1 | POST     "/api/create"
```

Old logs
```
[GIN] 2024/07/23 - 17:24:27 | 201 |  5.283626083s |       127.0.0.1 | POST     "/api/blobs/sha256:54696cbcadd1959275fc99f9cc67880d2f38419124da06cdf2140bad2dc3d94c"
[GIN] 2024/07/23 - 17:24:32 | 200 |  4.302020959s |       127.0.0.1 | POST     "/api/create"
```


resolves: https://github.com/ollama/ollama/issues/5388